# rapidapi/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Domain Intelligence API",
    "description": "Comprehensive domain intelligence in a single API call. Aggregate DNS records, SSL certificate details, WHOIS/RDAP registration data, subdomain enumeration (via Certificate Transparency logs), email security posture (SPF/DKIM/DMARC/BIMI), and ASN/geo info \u2014 all in one request or individually. Zero setup, no scraping headaches, JSON out of the box. Ideal for cybersecurity tools, lead enrichment, compliance checks, and SaaS apps.",
    "version": "1.0.0",
    "x-rapidapi-category": "Data",
    "contact": {
      "name": "Domain Intel API Support"
    }
  },
  "paths": {
    "/health": {
      "get": {
        "summary": "Health Check",
        "operationId": "health_health_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          }
        },
        "description": "Returns service status. No authentication required.",
        "tags": [
          "Utility"
        ]
      }
    },
    "/lookup": {
      "get": {
        "summary": "Full Domain Intelligence Lookup",
        "operationId": "lookup_lookup_get",
        "parameters": [
          {
            "name": "domain",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "default": "",
              "title": "Domain"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "description": "Aggregate endpoint: returns DNS, SSL, WHOIS, subdomains, email security, and ASN in one call. Most popular endpoint.",
        "tags": [
          "Aggregate"
        ]
      }
    },
    "/domain/{domain}/dns": {
      "get": {
        "summary": "DNS Records",
        "operationId": "dns_only_domain__domain__dns_get",
        "parameters": [
          {
            "name": "domain",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Domain"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "description": "Fetches all DNS record types (A, AAAA, MX, TXT, NS, CNAME, SOA) for the given domain.",
        "tags": [
          "DNS"
        ]
      }
    },
    "/domain/{domain}/ssl": {
      "get": {
        "summary": "SSL/TLS Certificate Details",
        "operationId": "ssl_only_domain__domain__ssl_get",
        "parameters": [
          {
            "name": "domain",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Domain"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "description": "Returns issuer, subject, validity dates, SANs, and cipher suite for the domain TLS certificate.",
        "tags": [
          "SSL"
        ]
      }
    },
    "/domain/{domain}/whois": {
      "get": {
        "summary": "WHOIS / RDAP Registration Data",
        "operationId": "whois_only_domain__domain__whois_get",
        "parameters": [
          {
            "name": "domain",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Domain"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "description": "Returns registrar, creation/expiry dates, registrant info, and nameservers via RDAP protocol.",
        "tags": [
          "WHOIS"
        ]
      }
    },
    "/domain/{domain}/subdomains": {
      "get": {
        "summary": "Subdomain Enumeration",
        "operationId": "subs_only_domain__domain__subdomains_get",
        "parameters": [
          {
            "name": "domain",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Domain"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "description": "Enumerates known subdomains using Certificate Transparency (crt.sh) logs. No brute force, instant results.",
        "tags": [
          "Subdomains"
        ]
      }
    },
    "/domain/{domain}/email-security": {
      "get": {
        "summary": "Email Security Posture",
        "operationId": "email_only_domain__domain__email_security_get",
        "parameters": [
          {
            "name": "domain",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Domain"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "description": "Checks SPF, DKIM (common selectors), DMARC, BIMI, and MTA-STS records. Useful for phishing risk assessment and compliance.",
        "tags": [
          "Email Security"
        ]
      }
    }
  },
  "components": {
    "schemas": {
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    }
  },
  "servers": [
    {
      "url": "https://87-99-152-31.sslip.io",
      "description": "Production"
    }
  ],
  "tags": [
    {
      "name": "Aggregate",
      "description": "All-in-one domain intelligence"
    },
    {
      "name": "DNS",
      "description": "DNS record lookups"
    },
    {
      "name": "SSL",
      "description": "TLS/SSL certificate inspection"
    },
    {
      "name": "WHOIS",
      "description": "Domain registration data"
    },
    {
      "name": "Subdomains",
      "description": "Subdomain discovery via CT logs"
    },
    {
      "name": "Email Security",
      "description": "SPF/DKIM/DMARC/BIMI checks"
    },
    {
      "name": "Utility",
      "description": "Service utilities"
    }
  ]
}